refactor: refactor import path
if we want to use shared libs in nx, you cannot set `paths` in tsconfig or it will confuse the tsc
so it won't locate your shared library. In order to use shared libs in the future, I deleted the
`paths` in recnet's tsconfig.json and update the import paths.